requirements.txt
asgiref==3.8.1
captcha==0.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
django-cors-headers==4.6.0
Django==5.1.3
djangorestframework==3.15.2
Faker==33.0.0
idna==3.10
djangorestframework-simplejwt==5.3.1
Pillow==11.0.0
PyJWT==2.10.0
requests==2.32.3
djangorestframework==3.15.2
django-ratelimit==4.1.0
six==1.16.0
sqlparse==0.5.2
typing-extensions==4.12.2
urllib3==2.2.3
Flask==3.1.0